!per #18360 Move Tagged to akka.persistence.journal
breaking api since 2.4.0-RC1, but only related to new things in 2.4 anyway
