+++ akka-docs/rst/scala/code/docs/persistence/PersistenceDocSpec.scala
@@ -8,7 +8,7 @@ import akka.actor._
 import akka.pattern.BackoffSupervisor
 import akka.persistence._
 import akka.stream.ActorMaterializer
-import akka.stream.scaladsl.{Source, Sink, Flow}
+import akka.stream.scaladsl.{ Source, Sink, Flow }
 
 import scala.concurrent.duration._
 import scala.language.postfixOps
